ensure AuthorizeHttpClientDecorator deletes cache on ForbiddenError

+++ lib/main/decorators/authorize_http_client_decorator.dart
@@ -24,11 +24,13 @@ class AuthorizeHttpClientDecorator implements HttpClient {
       final token = await fetchSecureCacheStorage.fetchSecure('token');
       final authorizedHeaders = headers ?? {}..addAll({'x-access-token': token});
       return await decoratee.request(url: url, method: method, body: body, headers: authorizedHeaders);
-    } on HttpError {
-      rethrow;
     } catch(error) {
-      await deleteSecureCacheStorage.deleteSecure('token');
-      throw HttpError.forbidden;
+      if (error is HttpError && error != HttpError.forbidden) {
+        rethrow;
+      } else {
+        await deleteSecureCacheStorage.deleteSecure('token');
+        throw HttpError.forbidden;
+      }
     }
   }
 }
